Add Qwen2-VL analysis scripts and visualization tools
- Implemented `32_generate_qwen2vl_visualizations.py` for generating visualizations comparing Qwen2-VL and PaLiGemma results.
- Created SLURM scripts for GPU (`slurm_30_qwen2vl_analysis_gpu.sh`) and CPU (`slurm_31_qwen2vl_analysis.sh`) analysis of Qwen2-VL.
- Added a testing script (`test_qwen2vl.py`) to validate model loading and inference.
- Included a SLURM script for testing (`slurm_test_qwen2vl.sh`) with GPU resources.

+++ sae_captioning_project/scripts/29_train_qwen2vl_sae.py
@@ -85,7 +85,7 @@ def load_qwen2vl_activations(path: str) -> tuple:
     """Load Qwen2-VL activation file."""
     data = torch.load(path, map_location='cpu', weights_only=False)
     
-    activations = data['activations']
+    activations = data["activations"].float()  # Convert bfloat16 to float32
     genders = data.get('genders', [])
     
     print(f"  Loaded: {activations.shape}")
